pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smart-expense-agent"
version = "1.0.0"
description = "AI-powered expense processing application with Streamlit and PostgreSQL"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Smart Expense Agent Team", email = "team@smartexpenseagent.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "streamlit==1.38.0",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.10",
    "groq>=0.31.0",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "pillow==10.4.0",
    "python-dotenv>=1.1.1",
    "numpy>=2.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/valonys/smart_agent"
Repository = "https://github.com/valonys/smart_agent.git"
Issues = "https://github.com/valonys/smart_agent/issues"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
